Expense Management System/requirements.txt
streamlit == 1.37.1
pandas == 2.2.3
pytest == 7.4.4
fastapi == 0.115.8
pydantic == 2.10.4
uvicorn == 0.34.0
mysql-connector-python == 9.2.0
requests == 2.32.3

